Use 'challenge' instead of 'nonce' + add temporary redirection
Update terminology to comply more with HOBA's spec (using "challenge" to
identify the one-time random string to be signed instead of "nonce").

Upon successful login, respond with a 300 pointing to the URL a user can
visit in their browser to obtain their session cookie. This allows
hpga-login(1) to cleanly interpret the response and open the pointed-to
url in the user's browser, thus removing the need for the user to copy a
url from their terminal into their browser.

**NOTE:** this redirection is ONLY for demo purposes. In fact, this
whole node web server shall be scrapped in favor of a PassportJS
auth strategy.